+++ src/internal/ajax/ajax.ts
@@ -1,6 +1,6 @@
 import { map } from '../operators/map';
 import { Observable } from '../Observable';
-import { AjaxConfig, AjaxRequest } from './types';
+import { AjaxConfig, AjaxRequest, AjaxDirection } from './types';
 import { AjaxResponse } from './AjaxResponse';
 import { AjaxTimeoutError, AjaxError } from './errors';
 
@@ -248,14 +248,14 @@ function ajaxGetJSON<T>(url: string, headers?: Record<string, string>): Observab
  * ```
  */
 export const ajax: AjaxCreationMethod = (() => {
-  const create = <T>(urlOrRequest: string | AjaxConfig) => {
-    const request =
-      typeof urlOrRequest === 'string'
+  const create = <T>(urlOrConfig: string | AjaxConfig) => {
+    const config: AjaxConfig =
+      typeof urlOrConfig === 'string'
         ? {
-            url: urlOrRequest,
+            url: urlOrConfig,
           }
-        : urlOrRequest;
-    return fromAjax<T>(request);
+        : urlOrConfig;
+    return fromAjax<T>(config);
   };
 
   create.get = ajaxGet;
@@ -268,6 +268,7 @@ export const ajax: AjaxCreationMethod = (() => {
   return create;
 })();
 
+const UPLOAD = 'upload';
 const DOWNLOAD = 'download';
 const LOADSTART = 'loadstart';
 const PROGRESS = 'progress';
@@ -275,21 +276,59 @@ const LOAD = 'load';
 
 export function fromAjax<T>(config: AjaxConfig): Observable<AjaxResponse<T>> {
   return new Observable((destination) => {
+    // Here we're pulling off each of the configuration arguments
+    // that we don't want to add to the request information we're
+    // passing around.
+    const { queryParams, body: configuredBody, headers: configuredHeaders, ...remainingConfig } = config;
+
+    let { url } = remainingConfig;
+    if (!url) {
+      throw new TypeError('url is required');
+    }
+
+    if (queryParams) {
+      let searchParams: URLSearchParams;
+      if (url.includes('?')) {
+        // If the user has passed a URL with a querystring already in it,
+        // we need to combine them. So we're going to split it. There
+        // should only be one `?` in a valid URL.
+        const parts = url.split('?');
+        if (2 < parts.length) {
+          throw new TypeError('invalid url');
+        }
+        // Add the passed queryParams to the params already in the url provided.
+        searchParams = new URLSearchParams(parts[1]);
+        // queryParams is converted to any because the runtime is *much* more permissive than
+        // the types are.
+        new URLSearchParams(queryParams as any).forEach((value, key) => searchParams.set(key, value));
+        // We have to do string concatenation here, because `new URL(url)` does
+        // not like relative URLs like `/this` without a base url, which we can't
+        // specify, nor can we assume `location` will exist, because of node.
+        url = parts[0] + '?' + searchParams;
+      } else {
+        // There is no pre-existing querystring, so we can just use URLSearchParams
+        // to convert the passed queryParams into the proper format and encodings.
+        // queryParams is converted to any because the runtime is *much* more permissive than
+        // the types are.
+        searchParams = new URLSearchParams(queryParams as any);
+        url = url + '?' + searchParams;
+      }
+    }
+
     // Normalize the headers. We're going to make them all lowercase, since
-    // Headers are case insenstive by design. This makes it easier to verify
+    // Headers are case insensitive by design. This makes it easier to verify
     // that we aren't setting or sending duplicates.
     const headers: Record<string, any> = {};
-    const requestHeaders = config.headers;
-    if (requestHeaders) {
-      for (const key in requestHeaders) {
-        if (requestHeaders.hasOwnProperty(key)) {
-          headers[key.toLowerCase()] = requestHeaders[key];
+    if (configuredHeaders) {
+      for (const key in configuredHeaders) {
+        if (configuredHeaders.hasOwnProperty(key)) {
+          headers[key.toLowerCase()] = configuredHeaders[key];
         }
       }
     }
 
     // Set the x-requested-with header. This is a non-standard header that has
-    // come to be a defacto standard for HTTP requests sent by libraries and frameworks
+    // come to be a de facto standard for HTTP requests sent by libraries and frameworks
     // using XHR. However, we DO NOT want to set this if it is a CORS request. This is
     // because sometimes this header can cause issues with CORS. To be clear,
     // None of this is necessary, it's only being set because it's "the thing libraries do"
@@ -301,8 +340,8 @@ export function fromAjax<T>(config: AjaxConfig): Observable<AjaxResponse<T>> {
 
     // Allow users to provide their XSRF cookie name and the name of a custom header to use to
     // send the cookie.
-    const { withCredentials, xsrfCookieName, xsrfHeaderName } = config;
-    if ((withCredentials || !config.crossDomain) && xsrfCookieName && xsrfHeaderName) {
+    const { withCredentials, xsrfCookieName, xsrfHeaderName } = remainingConfig;
+    if ((withCredentials || !remainingConfig.crossDomain) && xsrfCookieName && xsrfHeaderName) {
       const xsrfCookie = document?.cookie.match(new RegExp(`(^|;\\s*)(${xsrfCookieName})=([^;]*)`))?.pop() ?? '';
       if (xsrfCookie) {
         headers[xsrfHeaderName] = xsrfCookie;
@@ -311,7 +350,7 @@ export function fromAjax<T>(config: AjaxConfig): Observable<AjaxResponse<T>> {
 
     // Examine the body and determine whether or not to serialize it
     // and set the content-type in `headers`, if we're able.
-    const body = extractContentTypeAndMaybeSerializeBody(config.body, headers);
+    const body = extractContentTypeAndMaybeSerializeBody(configuredBody, headers);
 
     const _request: AjaxRequest = {
       // Default values
@@ -323,19 +362,15 @@ export function fromAjax<T>(config: AjaxConfig): Observable<AjaxResponse<T>> {
       responseType: 'json' as XMLHttpRequestResponseType,
 
       // Override with passed user values
-      ...config,
+      ...remainingConfig,
 
       // Set values we ensured above
+      url,
       headers,
       body,
     };
 
     let xhr: XMLHttpRequest;
-
-    const { url } = _request;
-    if (!url) {
-      throw new TypeError('url is required');
-    }
 
     // Create our XHR so we can get started.
     xhr = config.createXHR ? config.createXHR() : new XMLHttpRequest();
@@ -373,12 +408,12 @@ export function fromAjax<T>(config: AjaxConfig): Observable<AjaxResponse<T>> {
 
       /**
        * Creates a response object to emit to the consumer.
-       * @param direction the direction related to the event. Prefixes the event `type` in the response
-       * object. So "upload_" for events related to uploading, and "download_" for events related to
-       * downloading.
+       * @param direction the direction related to the event. Prefixes the event `type` in the
+       * `AjaxResponse` object with "upload_" for events related to uploading and "download_"
+       * for events related to downloading.
        * @param event the actual event object.
        */
-      const createResponse = (direction: 'upload' | 'download', event: ProgressEvent) =>
+      const createResponse = (direction: AjaxDirection, event: ProgressEvent) =>
         new AjaxResponse<T>(event, xhr, _request, `${direction}_${event.type}`);
 
       /**
@@ -391,14 +426,14 @@ export function fromAjax<T>(config: AjaxConfig): Observable<AjaxResponse<T>> {
        * @param direction The "direction", used to prefix the response object that is
        * emitted to the consumer. (e.g. "upload_" or "download_")
        */
-      const addProgressEvent = (target: any, type: string, direction: 'upload' | 'download') => {
+      const addProgressEvent = (target: any, type: string, direction: AjaxDirection) => {
         target.addEventListener(type, (event: ProgressEvent) => {
           destination.next(createResponse(direction, event));
         });
       };
 
       if (includeUploadProgress) {
-        [LOADSTART, PROGRESS, LOAD].forEach((type) => addProgressEvent(xhr.upload, type, 'upload'));
+        [LOADSTART, PROGRESS, LOAD].forEach((type) => addProgressEvent(xhr.upload, type, UPLOAD));
       }
 
       if (progressSubscriber) {
@@ -520,7 +555,7 @@ function extractContentTypeAndMaybeSerializeBody(body: any, headers: Record<stri
     // If we have made it here, this is an object, probably a POJO, and we'll try
     // to serialize it for them. If this doesn't work, it will throw, obviously, which
     // is okay. The workaround for users would be to manually set the body to their own
-    // serialized string (accounting for circular refrences or whatever), then set
+    // serialized string (accounting for circular references or whatever), then set
     // the content-type manually as well.
     headers['content-type'] = headers['content-type'] ?? 'application/json;charset=utf-8';
     return JSON.stringify(body);